requirements.txt
numpy
requests
scipy
librosa
unidecode
inflect
pyyaml
torchserve
torchserve-nightly 
torch-model-archiver-nightly 
torch-workflow-archiver-nightly